requirements.txt
asgiref==3.5.2
cfgv==3.3.1
crispy-bootstrap5==0.7
distlib==0.3.6
Django==3.2.16
django-autocomplete-light==3.9.4
django-colorfield==0.7.2
django-crispy-forms==1.14.0
django-htmx==1.12.2
django-tables2==2.4.1
filelock==3.8.0
identify==2.5.7
nodeenv==1.7.0
Pillow==9.2.0
platformdirs==2.5.2
pre-commit==2.20.0
pytz==2022.5
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
virtualenv==20.16.6
